# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Expert Quotes Requests Service",
    "description": "Bronze wrapper for journalist quote-request providers (Featured.com first; HARO/SOS/Qwoted later). Owns Featured.com auth lifecycle, rate-limit budgeting, cursor / sync state, and append-only bronze raw payloads.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3055"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PremiumQuestionsResponse": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PremiumQuestion"
            }
          },
          "refreshed": {
            "type": "boolean"
          }
        },
        "required": [
          "questions",
          "refreshed"
        ]
      },
      "PremiumQuestion": {
        "type": "object",
        "properties": {
          "featuredQuestionId": {
            "type": "integer"
          },
          "question": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "mediaOutlet": {
            "type": "string",
            "nullable": true
          },
          "pitchUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "featuredQuestionId",
          "question",
          "source",
          "mediaOutlet",
          "pitchUrl",
          "createdAt",
          "deadline"
        ]
      },
      "PremiumQuestionRefreshResponse": {
        "type": "object",
        "properties": {
          "refreshed": {
            "type": "boolean"
          },
          "inserted": {
            "type": "integer"
          },
          "updated": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          }
        },
        "required": [
          "refreshed",
          "inserted",
          "updated",
          "skipped"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/orgs/featured/premium-questions": {
      "get": {
        "summary": "List Featured premium questions (bronze-backed, lazy-refresh)",
        "description": "Lazy-refreshes from Featured `/premium-question-list` if last refresh > FEATURED_OPPORTUNITY_TTL_MS, persisting the raw payload to bronze and normalizing `mediaOutlet` through the shared outlet aliases. Returns the full currently-answerable premium set (no limit). `mediaOutlet` is `null` only when Featured exposes no outlet under any known key.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Premium questions + refresh flag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiumQuestionsResponse"
                }
              }
            }
          },
          "502": {
            "description": "Featured.com unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/featured/premium-questions/refresh": {
      "post": {
        "summary": "Force-refresh premium questions from Featured (bypass TTL)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Refresh stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiumQuestionRefreshResponse"
                }
              }
            }
          },
          "502": {
            "description": "Featured.com unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}